# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Maillume Hosted API",
    "version": "0.1.0-beta",
    "description": "Authenticated email risk assessment. Scan content and results are processed for the request and are not stored. Results are automated assessments, not guarantees."
  },
  "servers": [{ "url": "https://app.maillume.io" }],
  "paths": {
    "/api/v1/analyze": {
      "post": {
        "summary": "Analyze normalized email content",
        "security": [{ "apiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnalyzeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assessment completed",
            "headers": {
              "X-RateLimit-Limit": { "schema": { "type": "integer" } },
              "X-RateLimit-Remaining": { "schema": { "type": "integer" } }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalyzeResponse" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": { "type": "http", "scheme": "bearer", "bearerFormat": "mlm_..." }
    },
    "schemas": {
      "AnalyzeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["body"],
        "properties": {
          "source": { "type": "string", "enum": ["paste", "screenshot", "eml"] },
          "subject": { "type": "string", "maxLength": 300 },
          "senderEmail": { "type": "string", "maxLength": 320 },
          "locale": { "type": "string", "enum": ["en", "nl"], "default": "en" },
          "body": { "type": "string", "minLength": 1, "maxLength": 20000 },
          "links": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "string",
              "format": "uri",
              "pattern": "^[Hh][Tt][Tt][Pp][Ss]?://",
              "maxLength": 2048
            }
          },
          "linkPairs": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["displayedUrl", "destinationUrl"],
              "properties": {
                "displayedUrl": { "type": "string", "format": "uri", "pattern": "^[Hh][Tt][Tt][Pp][Ss]?://", "maxLength": 2048 },
                "destinationUrl": { "type": "string", "format": "uri", "pattern": "^[Hh][Tt][Tt][Pp][Ss]?://", "maxLength": 2048 }
              }
            }
          }
        }
      },
      "AnalysisResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["classification", "risk_level", "risk_score", "score_factors", "suspicious_signals", "detected_links", "recommended_action", "short_explanation"],
        "properties": {
          "classification": { "type": "string", "enum": ["likely_phishing", "likely_spam", "likely_legitimate", "uncertain"] },
          "risk_level": { "type": "string", "enum": ["low", "medium", "high"] },
          "risk_score": { "type": "integer", "minimum": 0, "maximum": 100, "description": "Versioned risk index equal to the sum of score_factors contributions; not a probability." },
          "score_factors": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["id", "family", "contribution", "label"],
              "properties": {
                "id": { "type": "string" },
                "family": { "type": "string", "enum": ["identity", "destination", "intent", "delivery", "style"] },
                "contribution": { "type": "integer", "minimum": 1, "maximum": 30 },
                "label": { "type": "string" }
              }
            }
          },
          "suspicious_signals": { "type": "array", "items": { "type": "string" } },
          "detected_links": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "recommended_action": { "type": "string" },
          "short_explanation": { "type": "string" }
        }
      },
      "AnalyzeResponse": {
        "type": "object",
        "required": ["result", "analysis_mode", "analysis_provider", "analysis_version", "disclaimer", "privacy"],
        "properties": {
          "result": { "$ref": "#/components/schemas/AnalysisResult" },
          "analysis_mode": { "type": "string", "enum": ["heuristic", "ai"] },
          "analysis_provider": { "type": "string", "enum": ["heuristic", "openai", "anthropic", "openai-compatible"] },
          "analysis_version": { "const": "analysis-v2.1" },
          "disclaimer": { "type": "string" },
          "privacy": {
            "type": "object",
            "required": ["stored", "retention", "message"],
            "properties": {
              "stored": { "const": false },
              "retention": { "const": "not_stored" },
              "message": { "type": "string" }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Request failed",
        "headers": {
          "Retry-After": { "schema": { "type": "integer" } }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  }
}